Página web de cliente, menu

+++ EntregaADomicilio.Core/Entidades/Platillo.cs
@@ -19,14 +19,11 @@ namespace EntregaADomicilio.Core.Entidades
 
         public string Categoria { get; set; }
 
-
         public string Nombre { get; set; }
-
 
         public string Descripcion { get; set; }
 
-
-        public double Precio { get; set; }
+        public decimal Precio { get; set; }
 
         public Archivo Archivo { get; set; }
     }
